% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_for_sequences.R
\name{adjustAlexandrovSignature}
\alias{adjustAlexandrovSignature}
\title{Adjust an Alexandrov signature (internal function)}
\usage{
adjustAlexandrovSignature(signature, factors)
}
\arguments{
\item{signature}{Single Alexandrov signature to be adjusted}

\item{factors}{Factors for trinucleotides}
}
\value{
Adjusted Alexandrov signature
}
\description{
Adjust an Alexandrov signature by a set of pattern-specific factors.
}
\references{
\url{http://rmpiro.net/decompTumor2Sig/}\cr
Krueger, Piro (2019) decompTumor2Sig: Identification of mutational
signature active in individual tumors. BMC Bioinformatics
20(Suppl 4):152.\cr
}
\author{
Rosario M. Piro\cr Politecnico di Milano\cr Maintainer: Rosario
M. Piro\cr E-Mail: <rmpiro@gmail.com> or <rosariomichael.piro@polimi.it>
}
\keyword{internal}
