% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_for_sequences.R
\name{adjustShiraishiSignature}
\alias{adjustShiraishiSignature}
\title{Adjust an Shiraishi signature (internal function)}
\usage{
adjustShiraishiSignature(signature, factorsMatrix)
}
\arguments{
\item{signature}{Single Shiraishi signature to be adjusted}

\item{factorsMatrix}{Factors for single nucleotides (Shiraishi-like matrix)}
}
\value{
Adjusted Shiraishi signature
}
\description{
Adjust an Shiraishi signature by a set of base-specific factors.
}
\references{
\url{http://rmpiro.net/decompTumor2Sig/}\cr
Krueger, Piro (2019) decompTumor2Sig: Identification of mutational
signature active in individual tumors. BMC Bioinformatics
20(Suppl 4):152.\cr
}
\author{
Rosario M. Piro\cr Politecnico di Milano\cr Maintainer: Rosario
M. Piro\cr E-Mail: <rmpiro@gmail.com> or <rosariomichael.piro@polimi.it>
}
\keyword{internal}
