% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildSortedAlexandrovSignaturePatternList.R
\name{buildSortedAlexandrovSignaturePatternList}
\alias{buildSortedAlexandrovSignaturePatternList}
\title{buildSortedAlexandrovSignaturePatternList (internal function)}
\usage{
buildSortedAlexandrovSignaturePatternList(numBases, trDir)
}
\arguments{
\item{numBases}{Number of bases for the sequence pattern (odd integer).}

\item{trDir}{Logical: use transcription-strand information?}
}
\value{
A sorted list of mutation features (e.g., triplets with base change).
}
\description{
Build and sort the pattern list of an Alexandrov signature in the following
way: First according to the base change: C>A, C>G, C>T, T>A, T>C, T>G,
then within these categories according to the flanking bases: A, C, G, T.
}
\references{
\url{http://rmpiro.net/decompTumor2Sig/}\cr
Krueger, Piro (2018) decompTumor2Sig: Identification of mutational
signatures active in individual tumors. BMC Bioinformatics (accepted for
publication).\cr
Krueger, Piro (2017) Identification of Mutational Signatures Active in
Individual Tumors. NETTAB 2017 - Methods, Tools & Platforms for
Personalized Medicine in the Big Data Era, October 16-18, 2017, Palermo,
Italy. PeerJ Preprints 5:e3257v1, 2017.
}
\author{
Rosario M. Piro\cr Freie Universitaet Berlin\cr Maintainer: Rosario
M. Piro\cr E-Mail: <rmpiro@gmail.com> or <r.piro@fu-berlin.de>
}
\keyword{internal}
