% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematical_functions.R
\name{computeFrobeniusNorm}
\alias{computeFrobeniusNorm}
\title{computeFrobeniusNorm (internal function)}
\usage{
computeFrobeniusNorm(A)
}
\arguments{
\item{A}{The numeric matrix.}
}
\value{
The Frobenius norm.
}
\description{
Compute the Frobenius norm of a numeric matrix (if this matrix is the 
difference between two matrices, then this corresponds to the Frobenius 
distance between these two matrices).
}
\references{
\url{http://rmpiro.net/decompTumor2Sig/}\cr
Krueger, Piro (2019) decompTumor2Sig: Identification of mutational
signatures active in individual tumors. BMC Bioinformatics
20(Suppl 4):152.\cr
}
\author{
Rosario M. Piro\cr Politecnico di Milano\cr Maintainer: Rosario
M. Piro\cr E-Mail: <rmpiro@gmail.com> or <rosariomichael.piro@polimi.it>
}
\keyword{internal}
