% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematical_functions.R
\name{computeRSS}
\alias{computeRSS}
\title{computeRSS (internal function)}
\usage{
computeRSS(x, y)
}
\arguments{
\item{x}{The first numeric object (e.g., matrix).}

\item{y}{The second numeric object (of the same type as \code{x}).}
}
\value{
The RSS.
}
\description{
Compute the residual sum of squares (RSS), i.e., the sum of squared errors.
}
\references{
\url{http://rmpiro.net/decompTumor2Sig/}\cr
Krueger, Piro (2017) Identification of Mutational Signatures Active in
Individual Tumors. NETTAB 2017 - Methods, Tools & Platforms for
Personalized Medicine in the Big Data Era, October 16-18, 2017, Palermo,
Italy. PeerJ Preprints 5:e3257v1, 2017.
}
\author{
Rosario M. Piro\cr Freie Universitaet Berlin\cr Maintainer: Rosario
M. Piro\cr E-Mail: <rmpiro@gmail.com> or <r.piro@fu-berlin.de>
}
\keyword{internal}
