% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convAlx2Shi.R
\name{convAlx2Shi}
\alias{convAlx2Shi}
\title{convAlx2Shi (internal function)}
\usage{
convAlx2Shi(x)
}
\arguments{
\item{x}{The Alexandrov signature (mutation pattern vector) to be converted.}
}
\value{
A Shiraishi signature (mutation pattern matrix).
}
\description{
Convert a single Alexandrov signature to a Shiraishi signature.
}
\references{
\url{http://rmpiro.net/decompTumor2Sig/}\cr
Krueger, Piro (2018) decompTumor2Sig: Identification of mutational
signatures active in individual tumors. BMC Bioinformatics (accepted for
publication).\cr
Krueger, Piro (2017) Identification of Mutational Signatures Active in
Individual Tumors. NETTAB 2017 - Methods, Tools & Platforms for
Personalized Medicine in the Big Data Era, October 16-18, 2017, Palermo,
Italy. PeerJ Preprints 5:e3257v1, 2017.
}
\author{
Rosario M. Piro\cr Freie Universitaet Berlin\cr Maintainer: Rosario
M. Piro\cr E-Mail: <rmpiro@gmail.com> or <r.piro@fu-berlin.de>
}
\keyword{internal}
