% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertMutTypesFromAlexandrov_v3_5.R
\name{convertMutTypesFromAlexandrov_v3_5}
\alias{convertMutTypesFromAlexandrov_v3_5}
\title{convertMutTypesFromAlexandrov_v3_5 (internal function)}
\usage{
convertMutTypesFromAlexandrov_v3_5(mutNames)
}
\arguments{
\item{mutNames}{A vector of mutation type names of format "NNNN"}
}
\value{
A vector of mutation type names of format "N[N>N]N"
}
\description{
Convert the mutation type naming from the "NNNN" to "N[N>N]N" (e.g.,
"ACGT" corresponds to "A[C>T]G"!)
}
\references{
\url{http://rmpiro.net/decompTumor2Sig/}\cr
Krueger, Piro (2019) decompTumor2Sig: Identification of mutational
signatures active in individual tumors. BMC Bioinformatics
20(Suppl 4):152.\cr
}
\author{
Rosario M. Piro\cr Politecnico di Milano\cr Maintainer: Rosario
M. Piro\cr E-Mail: <rmpiro@gmail.com> or <rosariomichael.piro@polimi.it>
}
\keyword{internal}
