% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_for_sequences.R
\name{convertSeqFreqToBaseFreq}
\alias{convertSeqFreqToBaseFreq}
\title{Convert sequence pattern frequencies to base frequencies (internal function)}
\usage{
convertSeqFreqToBaseFreq(frequencies, sigType)
}
\arguments{
\item{frequencies}{Vector for frequencies of individual sequence patterns}

\item{sigType}{Signature type information as returned by
\code{\link{determineTypeNumBasesAndTrDir}}}
}
\value{
Shiraishi signature-like matrix of per-base nucleotide frequencies
}
\description{
Convert sequence pattern frequencies (Alexandrov-like) to frequencies for
individual bases (Shiraishi-like).
}
\references{
\url{http://rmpiro.net/decompTumor2Sig/}\cr
Krueger, Piro (2019) decompTumor2Sig: Identification of mutational
signatures active in individual tumors. BMC Bioinformatics
20(Suppl 4):152.\cr
}
\author{
Rosario M. Piro\cr Politecnico di Milano\cr Maintainer: Rosario
M. Piro\cr E-Mail: <rmpiro@gmail.com> or <rosariomichael.piro@polimi.it>
}
\keyword{internal}
