% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determineTypeNumBasesAndTrDir.R
\name{determineTypeNumBasesAndTrDir}
\alias{determineTypeNumBasesAndTrDir}
\title{determineTypeNumBasesAndTrDir (internal function)}
\usage{
determineTypeNumBasesAndTrDir(mutData)
}
\arguments{
\item{mutData}{The mutation frequency data of the signature or genome.}
}
\value{
A list object composed of: "type"=type of signature or genome;
"numBases"=number of bases of the sequence patterns; and
"trDir"=logical value indicating whether transcription-strand information
has been considered in the mutation frequency data.
}
\description{
For a given signature or genome representation: determine the
type (Shiraishi or Alexandrov), the number of bases, and whether
transcription-strand information is included.
}
\references{
\url{http://rmpiro.net/decompTumor2Sig/}\cr
Krueger, Piro (2018) decompTumor2Sig: Identification of mutational
signatures active in individual tumors. BMC Bioinformatics (accepted for
publication).\cr
Krueger, Piro (2017) Identification of Mutational Signatures Active in
Individual Tumors. NETTAB 2017 - Methods, Tools & Platforms for
Personalized Medicine in the Big Data Era, October 16-18, 2017, Palermo,
Italy. PeerJ Preprints 5:e3257v1, 2017.
}
\author{
Rosario M. Piro\cr Freie Universitaet Berlin\cr Maintainer: Rosario
M. Piro\cr E-Mail: <rmpiro@gmail.com> or <r.piro@fu-berlin.de>
}
\keyword{internal}
