% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_classes_wrapper_functions.R
\name{getCountData}
\alias{getCountData}
\title{getCountData (internal function)}
\usage{
getCountData(x)
}
\arguments{
\item{x}{Object from which to get the slot data.}
}
\value{
Slot data.
}
\description{
`getCountData()` extracts the content of the \code{countData}
slot (numeric values) from a \code{pmsignature::MutationFeatureData} object.
}
\references{
\url{http://rmpiro.net/decompTumor2Sig/}\cr
Krueger, Piro (2019) decompTumor2Sig: Identification of mutational
signatures active in individual tumors. BMC Bioinformatics
20(Suppl 4):152.\cr
}
\author{
Rosario M. Piro\cr Politecnico di Milano\cr Maintainer: Rosario
M. Piro\cr E-Mail: <rmpiro@gmail.com> or <rosariomichael.piro@polimi.it>
}
\keyword{internal}
