% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_classes_wrapper_functions.R
\name{getSigFromEstParam}
\alias{getSigFromEstParam}
\title{getSigFromEstParam (internal function)}
\usage{
getSigFromEstParam(x, signum)
}
\arguments{
\item{x}{Object from which to get the signature.}

\item{signum}{Number (1..N) of the signature to be extracted.}
}
\value{
Signature.
}
\description{
`getSigFromEstParam()` extracts a specific signature from the
\code{signatureFeatureDistribution} slot of a
\code{pmsignature::EstimatedParameters} object.
}
\references{
\url{http://rmpiro.net/decompTumor2Sig/}\cr
Krueger, Piro (2019) decompTumor2Sig: Identification of mutational
signatures active in individual tumors. BMC Bioinformatics
20(Suppl 4):152.\cr
}
\author{
Rosario M. Piro\cr Politecnico di Milano\cr Maintainer: Rosario
M. Piro\cr E-Mail: <rmpiro@gmail.com> or <rosariomichael.piro@polimi.it>
}
\keyword{internal}
