% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_classes_wrapper_functions.R
\name{isEstParamObject}
\alias{isEstParamObject}
\title{isEstParamObject (internal function)}
\usage{
isEstParamObject(x)
}
\arguments{
\item{x}{Object for which to verify compatibility.}
}
\value{
Logical value (true or false).
}
\description{
`isEstParamObject()` checks whether an object is compatible to a
\code{pmsignature::EstimatedParameters} object (version 0.3.0), i.e.,
whether it contains the same info in the same format.
}
\references{
\url{http://rmpiro.net/decompTumor2Sig/}\cr
Krueger, Piro (2018) decompTumor2Sig: Identification of mutational
signatures active in individual tumors. BMC Bioinformatics (accepted for
publication).\cr
Krueger, Piro (2017) Identification of Mutational Signatures Active in
Individual Tumors. NETTAB 2017 - Methods, Tools & Platforms for
Personalized Medicine in the Big Data Era, October 16-18, 2017, Palermo,
Italy. PeerJ Preprints 5:e3257v1, 2017.
}
\author{
Rosario M. Piro\cr Freie Universitaet Berlin\cr Maintainer: Rosario
M. Piro\cr E-Mail: <rmpiro@gmail.com> or <r.piro@fu-berlin.de>
}
\keyword{internal}
