% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_classes_wrapper_functions.R
\name{isMutFeatDataObject}
\alias{isMutFeatDataObject}
\title{isMutFeatDataObject (internal function)}
\usage{
isMutFeatDataObject(x)
}
\arguments{
\item{x}{Object for which to verify compatibility.}
}
\value{
Logical value (true or false).
}
\description{
`isMutFeatDataObject()` checks whether an object is compatible to a
\code{pmsignature::MutationFeatureData} object (version 0.3.0), i.e.,
whether it contains the same info in the same format.
}
\references{
\url{http://rmpiro.net/decompTumor2Sig/}\cr
Krueger, Piro (2019) decompTumor2Sig: Identification of mutational
signatures active in individual tumors. BMC Bioinformatics
20(Suppl 4):152.\cr
}
\author{
Rosario M. Piro\cr Politecnico di Milano\cr Maintainer: Rosario
M. Piro\cr E-Mail: <rmpiro@gmail.com> or <rosariomichael.piro@polimi.it>
}
\keyword{internal}
