% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotShiraishiModel.R
\name{plotShiraishiModel}
\alias{plotShiraishiModel}
\title{plotShiraishiModel (internal function)}
\usage{
plotShiraishiModel(mutData, numBases, trDir, colors = NULL)
}
\arguments{
\item{mutData}{(Mandatory) The signature or genome mutation frequency data
to be plotted.}

\item{numBases}{The number of bases of the sequence pattern.}

\item{trDir}{Logical value specifying whether transcription strand
information is present.}

\item{colors}{Vector of four colors to be used for the base data.
If \code{NULL} (default), the colors are consensus colors used for
sequence logos.}
}
\value{
Returns (or draws) a plot according to the Shiraishi model of
mutational signatures.
}
\description{
`plotShiraishiModel()` plots a single signature or the mutation
frequency data for a single genome of the Shiraishi-type model.
}
\references{
\url{http://rmpiro.net/decompTumor2Sig/}\cr
Krueger, Piro (2019) decompTumor2Sig: Identification of mutational
signatures active in individual tumors. BMC Bioinformatics
20(Suppl 4):152.\cr
}
\author{
Rosario M. Piro\cr Politecnico di Milano\cr Maintainer: Rosario
M. Piro\cr E-Mail: <rmpiro@gmail.com> or <rosariomichael.piro@polimi.it>
}
\keyword{internal}
