% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setNames4ShiraishiTable.R
\name{setNames4ShiraishiTable}
\alias{setNames4ShiraishiTable}
\title{setNames4ShiraishiTable (internal function)}
\usage{
setNames4ShiraishiTable(table)
}
\arguments{
\item{table}{The table to be named.}
}
\value{
The same table with named columns and rows.
}
\description{
Set row and column names of a Shiraishi-model genome or signature table.
}
\references{
\url{http://rmpiro.net/decompTumor2Sig/}\cr
Krueger, Piro (2018) decompTumor2Sig: Identification of mutational
signatures active in individual tumors. BMC Bioinformatics (accepted for
publication).\cr
Krueger, Piro (2017) Identification of Mutational Signatures Active in
Individual Tumors. NETTAB 2017 - Methods, Tools & Platforms for
Personalized Medicine in the Big Data Era, October 16-18, 2017, Palermo,
Italy. PeerJ Preprints 5:e3257v1, 2017.
}
\author{
Rosario M. Piro\cr Freie Universitaet Berlin\cr Maintainer: Rosario
M. Piro\cr E-Mail: <rmpiro@gmail.com> or <r.piro@fu-berlin.de>
}
\keyword{internal}
