% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_classes_wrapper_functions.R
\name{setTrDir}
\alias{setTrDir}
\title{setTrDir (internal function)}
\usage{
setTrDir(x, value)
}
\arguments{
\item{x}{Object of which to change slot data.}

\item{value}{Value to be written to the slot.}
}
\value{
The modified object \code{x}.
}
\description{
`setTrDir()` serves as wrapper functions to change content of the
\code{transcriptionDirection} slot (logical value) of a
\code{pmsignature::MutationFeatureData} object.
}
\references{
\url{http://rmpiro.net/decompTumor2Sig/}\cr
Krueger, Piro (2019) decompTumor2Sig: Identification of mutational
signatures active in individual tumors. BMC Bioinformatics
20(Suppl 4):152.\cr
}
\author{
Rosario M. Piro\cr Politecnico di Milano\cr Maintainer: Rosario
M. Piro\cr E-Mail: <rmpiro@gmail.com> or <rosariomichael.piro@polimi.it>
}
\keyword{internal}
