% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematical_functions.R
\name{computeRSS}
\alias{computeRSS}
\title{computeRSS (internal function)}
\usage{
computeRSS(x, y)
}
\arguments{
\item{x}{The first numeric object (e.g., matrix).}

\item{y}{The second numeric object (of the same type as \code{x}).}
}
\value{
The RSS.
}
\description{
Compute the residual sum of squares (RSS), i.e., the sum of squared errors.
}
\references{
\url{http://rmpiro.net/decompTumor2Sig/}\cr
Krueger, Piro (2019) decompTumor2Sig: Identification of mutational
signatures active in individual tumors. BMC Bioinformatics
20(Suppl 4):152.\cr
}
\author{
Rosario M. Piro\cr Politecnico di Milano\cr Maintainer: Rosario
M. Piro\cr E-Mail: <rmpiro@gmail.com> or <rosariomichael.piro@polimi.it>
}
\keyword{internal}
