% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convAlx2Shi.R
\name{convAlx2Shi}
\alias{convAlx2Shi}
\title{convAlx2Shi (internal function)}
\usage{
convAlx2Shi(x)
}
\arguments{
\item{x}{The Alexandrov signature (mutation pattern vector) to be converted.}
}
\value{
A Shiraishi signature (mutation pattern matrix).
}
\description{
Convert a single Alexandrov signature to a Shiraishi signature.
}
\references{
\url{http://rmpiro.net/decompTumor2Sig/}\cr
Krueger, Piro (2019) decompTumor2Sig: Identification of mutational
signatures active in individual tumors. BMC Bioinformatics 
20(Suppl 4):152.\cr
}
\author{
Rosario M. Piro\cr Politecnico di Milano\cr Maintainer: Rosario
M. Piro\cr E-Mail: <rmpiro@gmail.com> or <rosariomichael.piro@polimi.it>
}
\keyword{internal}
