% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertAlexandrov2Shiraishi.R
\name{convertAlexandrov2Shiraishi}
\alias{convertAlexandrov2Shiraishi}
\title{Convert Alexandrov-type signatures to Shiraishi signatures}
\usage{
convertAlexandrov2Shiraishi(signatures)
}
\arguments{
\item{signatures}{(Mandatory) A list of Alexandrov signatures with named
elements as produced by \code{readAlexandrovSignatures}.}
}
\value{
A list of Shiraishi signatures that can be used for
\code{decomposeTumorGenomes}.
}
\description{
`convertAlexandrov2Shiraishi()` converts a set Alexandrov signatures to the
Shiraishi model of signatures, summing the respective frequencies of base
changes, and upstream and downstream flanking bases. The resulting
Shiraishi signatures don't provide information on the transcription strand,
as this is not part of the Alexandrov signatures. [Attention: this
conversion entails a loss of specificity and the applicability of Shiraishi
signatures derived from Alexandrov signatures has not been extensively
explored!]
}
\examples{

### get Alexandrov signatures from COSMIC
signAlexandrov <- readAlexandrovSignatures()

### convert them to the Shiraishi model
signShiraishi <- convertAlexandrov2Shiraishi(signAlexandrov)

}
\references{
\url{http://rmpiro.net/decompTumor2Sig/}\cr
Krueger, Piro (2017) Identification of Mutational Signatures Active in
Individual Tumors. NETTAB 2017 - Methods, Tools & Platforms for
Personalized Medicine in the Big Data Era, October 16-18, 2017, Palermo,
Italy. PeerJ Preprints 5:e3257v1, 2017.
}
\seealso{
\code{\link{decompTumor2Sig}}\cr
\code{\link{readAlexandrovSignatures}}\cr
\code{\link{readShiraishiSignatures}}
}
\author{
Rosario M. Piro\cr Freie Universitaet Berlin\cr
Maintainer: Rosario M. Piro\cr E-Mail: <r.piro@fu-berlin.de> or
<rmpiro@gmail.com>
}
