% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_classes_wrapper_functions.R
\name{getSigFeatDist}
\alias{getSigFeatDist}
\title{getSigFeatDist (internal function)}
\usage{
getSigFeatDist(x)
}
\arguments{
\item{x}{Object from which to get the slot data.}
}
\value{
Slot data.
}
\description{
getSigFeatDist`()` extracts the content of the
\code{signatureFeatureDistribution} slot (string) from a
\code{pmsignature::EstimatedParameters} object.
}
\references{
\url{http://rmpiro.net/decompTumor2Sig/}\cr
Krueger, Piro (2017) Identification of Mutational Signatures Active in
Individual Tumors. NETTAB 2017 - Methods, Tools & Platforms for
Personalized Medicine in the Big Data Era, October 16-18, 2017, Palermo,
Italy. PeerJ Preprints 5:e3257v1, 2017.
}
\author{
Rosario M. Piro\cr Freie Universitaet Berlin\cr Maintainer: Rosario
M. Piro\cr E-Mail: <rmpiro@gmail.com> or <r.piro@fu-berlin.de>
}
\keyword{internal}
