% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSignaturesFromEstParam.R
\name{getSignaturesFromEstParam}
\alias{getSignaturesFromEstParam}
\title{Get signatures from an \code{EstimatedParameters} object.}
\usage{
getSignaturesFromEstParam(Param)
}
\arguments{
\item{Param}{(Mandatory) A \code{pmsignature::EstimatedParameters} object
as those produced by the de novo signature construction method
\code{pmsignature::getPMSignature}.}
}
\value{
A list of Shiraishi signatures, one object per signature. Please
see \code{readShiraishiSignatures} or the \code{decompTumor2Sig} vignette
for more information on the format of Shiraishi signatures.
}
\description{
`getSignaturesFromEstParam()` takes an \code{EstimatedParameters} object
(signatures data) as computed by the '\code{pmsignature}' package (by
\code{pmsignature::getPMSignature}; version 0.3.0) and extracts the
signature information. The signatures can then be passed to
\code{decomposeTumorGenomes}.
}
\examples{

### load signatures for breast cancer genomes from 
### Nik-Zainal et al (PMID: 22608084) in the format produced by
### pmsignature (PMID: 26630308)
pmsigdata <- system.file("extdata",
         "Nik-Zainal_PMID_22608084-pmsignature-Param.Rdata", 
         package="decompTumor2Sig")
load(pmsigdata)

### extract the signatures from the pmsignature Param object
signatures <- getSignaturesFromEstParam(Param)

}
\references{
\url{http://rmpiro.net/decompTumor2Sig/}\cr
Krueger, Piro (2018) decompTumor2Sig: Identification of mutational
signatures active in individual tumors. BMC Bioinformatics (accepted for
publication).\cr
Krueger, Piro (2017) Identification of Mutational Signatures Active in
Individual Tumors. NETTAB 2017 - Methods, Tools & Platforms for
Personalized Medicine in the Big Data Era, October 16-18, 2017, Palermo,
Italy. PeerJ Preprints 5:e3257v1, 2017.
}
\seealso{
\code{\link{decompTumor2Sig}}\cr
\code{\link{readShiraishiSignatures}}
}
\author{
Rosario M. Piro and Sandra Krueger\cr Freie Universitaet Berlin\cr
Maintainer: Rosario M. Piro\cr E-Mail: <rmpiro@gmail.com> or
<r.piro@fu-berlin.de>
}
