% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_for_class_checks.R
\name{isSignatureSet}
\alias{isSignatureSet}
\title{isSignatureSet}
\usage{
isSignatureSet(x)
}
\arguments{
\item{x}{Object to be checked.}
}
\value{
Logical value (true or false).
}
\description{
`isSignatureSet()` checks whether the input object is a set (list) of
numeric objects compatible with either the Alexandrov format (probability
vectors; see \code{isAlexandrovSet}) or the Shiraishi format (matrices or
data.frames of probabilities; see \code{isShiraishiSet}).
NOTE: These can also be genomes compatible with one of the two formats!
}
\examples{

### get Alexandrov signatures from COSMIC
signAlexandrov <- readAlexandrovSignatures()

isSignatureSet(signAlexandrov)
}
\references{
\url{http://rmpiro.net/decompTumor2Sig/}\cr
Krueger, Piro (2018) decompTumor2Sig: Identification of mutational
signatures active in individual tumors. BMC Bioinformatics (accepted for
publication).\cr
Krueger, Piro (2017) Identification of Mutational Signatures Active in
Individual Tumors. NETTAB 2017 - Methods, Tools & Platforms for
Personalized Medicine in the Big Data Era, October 16-18, 2017, Palermo,
Italy. PeerJ Preprints 5:e3257v1, 2017.
}
\seealso{
\code{\link{decompTumor2Sig}}\cr
\code{\link{isAlexandrovSet}}\cr
\code{\link{isShiraishiSet}}
}
\author{
Rosario M. Piro\cr Freie Universitaet Berlin\cr Maintainer: Rosario
M. Piro\cr E-Mail: <rmpiro@gmail.com> or <r.piro@fu-berlin.de>
}
