% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAlexandrovSignatures.R
\name{readAlexandrovSignatures}
\alias{readAlexandrovSignatures}
\title{Read Alexandrov-type signatures (COSMIC format).}
\usage{
readAlexandrovSignatures(file)
}
\arguments{
\item{file}{(Mandatory) Can be a file name or an URL for download.
Default (COSMIC):
"https://cancer.sanger.ac.uk/cancergenome/assets/signatures_probabilities.txt"}
}
\value{
A list of Alexandrov signatures that can be used for
\code{decomposeTumorGenomes}.
}
\description{
`readAlexandrovSignatures()` reads a set of Alexandrov-type signatures
(COSMIC format) from a flat file or URL. Signatures must be specified in the
tab-separated format used by the COSMIC website; see details below or\cr
\url{http://cancer.sanger.ac.uk/cosmic/signatures} -> "Download signatures".
}
\details{
COSMIC format for Alexandrov signatures:

\tabular{llllll}{
Subst. \tab Trinucleotide \tab Mutation Type \tab Signature 1 \tab
Signature 2 \tab...\cr
C>A \tab ACA \tab A[C>A]A \tab 0.0110983262 \tab 0.0006827082 \tab...\cr
C>A \tab ACC \tab A[C>A]C \tab 0.0091493407 \tab 0.0006191072 \tab...\cr
C>A \tab ACG \tab A[C>A]G \tab 0.0014900705 \tab 0.0000992790 \tab...\cr
C>A \tab ACT \tab A[C>A]T \tab 0.0062338852 \tab 0.0003238914 \tab...\cr
[...]\cr
T>G \tab TTG \tab T[T>G]G \tab 0.0020310769 \tab 0.0002066152 \tab...\cr
T>G \tab TTT \tab T[T>G]T \tab 0.0040301281 \tab 0.0000235982 \tab...\cr
}
}
\examples{

### get Alexandrov signatures from COSMIC
signatures <- readAlexandrovSignatures()

}
\references{
\url{http://rmpiro.net/decompTumor2Sig/}\cr
Krueger, Piro (2018) decompTumor2Sig: Identification of mutational
signatures active in individual tumors. BMC Bioinformatics (accepted for
publication).\cr
Krueger, Piro (2017) Identification of Mutational Signatures Active in
Individual Tumors. NETTAB 2017 - Methods, Tools & Platforms for
Personalized Medicine in the Big Data Era, October 16-18, 2017, Palermo,
Italy. PeerJ Preprints 5:e3257v1, 2017.
}
\seealso{
\code{\link{decompTumor2Sig}}\cr
\code{\link{readShiraishiSignatures}}
}
\author{
Rosario M. Piro\cr Freie Universitaet Berlin\cr Maintainer: Rosario
M. Piro\cr E-Mail: <rmpiro@gmail.com> or <r.piro@fu-berlin.de>
}
