% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setNames4ShiraishiTable.R
\name{setNames4ShiraishiTable}
\alias{setNames4ShiraishiTable}
\title{setNames4ShiraishiTable (internal function)}
\usage{
setNames4ShiraishiTable(table)
}
\arguments{
\item{table}{The table to be named.}
}
\value{
The same table with named columns and rows.
}
\description{
Set row and column names of a Shiraishi-model genome or signature table.
}
\references{
\url{http://rmpiro.net/decompTumor2Sig/}\cr
Krueger, Piro (2019) decompTumor2Sig: Identification of mutational
signatures active in individual tumors. BMC Bioinformatics
20(Suppl 4):152.\cr
}
\author{
Rosario M. Piro\cr Politecnico di Milano\cr Maintainer: Rosario
M. Piro\cr E-Mail: <rmpiro@gmail.com> or <rosariomichael.piro@polimi.it>
}
\keyword{internal}
